Revert Suspense header/footer streaming (worsened FCP by 0.3s)
Suspense boundaries on Header/Footer added streaming overhead that
consistently increased FCP from 1.4s to 1.7s across 4 mobile runs.
Restore direct category fetching in storefront layout, remove
StorefrontShell, and revert CartButton mounted guard (no longer
needed without Suspense-induced hydration mismatch).
