+++ src/components/layout/CartButton.tsx
@@ -2,18 +2,12 @@
 
 import { ShoppingBag } from "lucide-react";
 import { useTranslations } from "next-intl";
-import { useEffect, useState } from "react";
 import { Button } from "@/components/ui/button";
 import { useCart } from "@/contexts/CartContext";
 
 export function CartButton() {
   const t = useTranslations("header");
   const { itemCount, openCart } = useCart();
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => {
-    setMounted(true);
-  }, []);
 
   return (
     <Button
@@ -24,7 +18,7 @@ export function CartButton() {
       className="relative"
     >
       <ShoppingBag className="size-5" />
-      {mounted && itemCount > 0 && (
+      {itemCount > 0 && (
         <span className="absolute top-0 right-0 bg-primary text-white text-xs font-medium rounded-full h-5 w-5 flex items-center justify-center">
           {itemCount}
         </span>
